requirements.txt
pywin32
psutil
requests
python-dotenv
